requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==25.1.0
    # via -r requirements.in
alembic==1.18.4
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via starlette
argon2-cffi==25.1.0
    # via -r requirements.in
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
asyncpg==0.31.0
    # via -r requirements.in
authlib==1.7.0
    # via -r requirements.in
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
click==8.3.3
    # via uvicorn
cryptography==47.0.0
    # via
    #   authlib
    #   joserfc
dotenv==0.9.9
    # via -r requirements.in
fastapi==0.136.1
    # via -r requirements.in
greenlet==3.5.0
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.13
    # via anyio
iniconfig==2.3.0
    # via pytest
joserfc==1.6.4
    # via authlib
librt==0.9.0
    # via mypy
mako==1.3.12
    # via alembic
markupsafe==3.0.3
    # via mako
mypy==1.20.2
    # via -r requirements.in
mypy-extensions==1.1.0
    # via mypy
packaging==26.2
    # via pytest
passlib==1.7.4
    # via -r requirements.in
pathspec==1.1.1
    # via mypy
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.12
    # via -r requirements.in
pycparser==3.0
    # via cffi
pydantic==2.13.3
    # via fastapi
pydantic-core==2.46.3
    # via pydantic
pygments==2.20.0
    # via pytest
pyjwt==2.12.1
    # via -r requirements.in
pytest==9.0.3
    # via -r requirements.in
python-dotenv==1.2.2
    # via dotenv
python-multipart==0.0.27
    # via -r requirements.in
ruff==0.15.12
    # via -r requirements.in
sqlalchemy==2.0.49
    # via
    #   -r requirements.in
    #   alembic
starlette==1.0.0
    # via fastapi
types-aiofiles==25.1.0.20260409
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
uvicorn==0.46.0
    # via -r requirements.in
